Create ProjectDriver.java
The final project for Object Oriented Programming with Prog Boustique. While I made had gotten a C in this class, it was due to testing poorly because I struggle with code tracing due to my ADHD and missing small details. All of projects worked and all of them in this course got As